inputable/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
anthropic==0.40.0
openai==1.57.0
genanki==0.13.0
httpx==0.28.1
pillow==11.0.0
pytest==8.3.4
pytest-asyncio==0.24.0
hypothesis==6.122.3
